test clipsize inc 4, from iter 7, dec 0.2, only on ripup 0

+++ src/drt/src/dr/FlexDR.h
@@ -151,7 +151,7 @@ class FlexDR
   std::string dist_dir_;
   std::string globals_path_;
   bool increaseClipsize_;
-  int clipSizeInc_;
+  float clipSizeInc_;
 
   // others
   void init();
